fix(mcp): share downsampleSeries helper and clamp safeLimit defaults
